Read Me
In this project, we implement a prototype of a Quantum Neural Network for the Iris dataset ( available at Scikit-Learn web-site) using Qiskit and test it on a real quantum computer provided by IBM-Quantum Experience. The data set has 150 samples, 50 samples for each type. Each sample has four features, namely: Sepal Length, Sepal Width, Petal Length and Petal Width. In this experiment, however, we decided to do binary classification. Therefore, we only take the first 100 samples for the first two types of irises(Setosa and Versicolour). We divide the dataset into two sets, with 67 samples in the training test (X_train), and 33 samples in the testing set (X_test). We found that the QNN learn very fast, but the accuracy fluctuates once it reaches 80% due to the error in the qubits.